Fix the test with timers.
This should resolve #7.
According to my analysis the fsm_timers test used to fail sometimes because this
test's assertions assume strict realtime timers. The main issue with the
previous test is that it is doing `sleep` in order to allow multiple timers to
fire. `sleep` doesn't guarantee that the process will be sleeping for the exact
specified duration, it only guarantees that it will sleep for at least the
requested duration. This means that when sleeping for 450ms we used to expect
that there will be 4 firings of the 100ms timer, but in reality there could be
more is the process sleeps for >500ms.
Suspicion: I think in reality it is a low chance for a process to sleep extra
50ms. This seems to be more likely on macos runners (because they are more
oversubscribed and overloaded?). To have a better chance of reproducing the
issue I reduced timer/sleep durations in the test 10x.
I couldn't find a way to refactor the test due to the way timers work in finny,
i.e. timer events do not get dispatched until you call `dispatch_timer_events`,
which means timers firing do not actually map to the timestamps one would
expect. Hence, the test is now trying to see how long the process actually was
sleeping for, and calculates the expected number of timers (and light toggles)
based on that.